+++ tubarr/config.py
@@ -1,0 +1,119 @@
+import os
+import yaml
+import logging
+from typing import Dict
+
+logger = logging.getLogger("yt-to-jellyfin")
+
+
+def _load_config() -> Dict:
+    """Load configuration from environment variables or config file."""
+    # Check for a local yt-dlp in the same directory as this file
+    script_dir = os.path.dirname(os.path.abspath(__file__))
+    local_ytdlp = os.path.join(script_dir, "yt-dlp")
+    if os.path.exists(local_ytdlp) and os.access(local_ytdlp, os.X_OK):
+        ytdlp_default = local_ytdlp
+    else:
+        specific_paths = [
+            "/home/marc/Documents/yt-dlp/yt-dlp",
+            "/usr/local/bin/yt-dlp",
+            "/usr/bin/yt-dlp",
+        ]
+        for path in specific_paths:
+            if os.path.exists(path) and os.access(path, os.X_OK):
+                ytdlp_default = path
+                break
+        else:
+            ytdlp_default = "yt-dlp"
+
+    config = {
+        "output_dir": os.environ.get("OUTPUT_DIR", "./media"),
+        "quality": os.environ.get("VIDEO_QUALITY", "1080"),
+        "use_h265": os.environ.get("USE_H265", "true").lower() == "true",
+        "crf": int(os.environ.get("CRF", "28")),
+        "ytdlp_path": os.environ.get("YTDLP_PATH", ytdlp_default),
+        "cookies": "",
+        "completed_jobs_limit": int(os.environ.get("COMPLETED_JOBS_LIMIT", "10")),
+        "web_enabled": os.environ.get("WEB_ENABLED", "true").lower() == "true",
+        "web_port": int(os.environ.get("WEB_PORT", "8000")),
+        "web_host": os.environ.get("WEB_HOST", "0.0.0.0"),
+        "update_checker_enabled": os.environ.get("UPDATE_CHECKER_ENABLED", "false").lower() == "true",
+        "update_checker_interval": int(os.environ.get("UPDATE_CHECKER_INTERVAL", "60")),
+        "jellyfin_enabled": os.environ.get("JELLYFIN_ENABLED", "false").lower() == "true",
+        "jellyfin_tv_path": os.environ.get("JELLYFIN_TV_PATH", ""),
+        "jellyfin_host": os.environ.get("JELLYFIN_HOST", ""),
+        "jellyfin_port": os.environ.get("JELLYFIN_PORT", "8096"),
+        "jellyfin_api_key": os.environ.get("JELLYFIN_API_KEY", ""),
+        "clean_filenames": os.environ.get("CLEAN_FILENAMES", "true").lower() == "true",
+    }
+
+    cookies_path = os.environ.get("COOKIES_PATH", "")
+    if cookies_path and os.path.exists(cookies_path):
+        config["cookies"] = cookies_path
+    elif cookies_path:
+        logger.warning(f"Cookies file not found at {cookies_path}, ignoring")
+
+    config_file = os.environ.get("CONFIG_FILE", "config/config.yml")
+    if os.path.exists(config_file):
+        try:
+            with open(config_file, "r") as f:
+                file_config = yaml.safe_load(f)
+            if file_config and isinstance(file_config, dict):
+                if "media" in file_config and isinstance(file_config["media"], dict):
+                    for key, value in file_config["media"].items():
+                        if key == "output_dir":
+                            config["output_dir"] = value
+                        elif key == "quality":
+                            config["quality"] = value
+                        elif key == "use_h265":
+                            config["use_h265"] = value
+                        elif key == "crf":
+                            config["crf"] = int(value)
+                        elif key == "clean_filenames":
+                            config["clean_filenames"] = value
+
+                if "cookies_path" in file_config:
+                    cookies_path = file_config["cookies_path"]
+                    if os.path.exists(cookies_path):
+                        config["cookies"] = cookies_path
+                    else:
+                        logger.warning(f"Cookies file not found at {cookies_path}, ignoring")
+
+                if "defaults" in file_config and isinstance(file_config["defaults"], dict):
+                    config["defaults"] = file_config["defaults"]
+
+                if "web" in file_config and isinstance(file_config["web"], dict):
+                    for key, value in file_config["web"].items():
+                        if key == "enabled":
+                            config["web_enabled"] = value
+                        elif key == "port":
+                            config["web_port"] = int(value)
+                        elif key == "host":
+                            config["web_host"] = value
+
+                if "jellyfin" in file_config and isinstance(file_config["jellyfin"], dict):
+                    for key, value in file_config["jellyfin"].items():
+                        if key == "enabled":
+                            config["jellyfin_enabled"] = value
+                        elif key == "tv_path":
+                            config["jellyfin_tv_path"] = value
+                        elif key == "host":
+                            config["jellyfin_host"] = value
+                        elif key == "port":
+                            config["jellyfin_port"] = str(value)
+                        elif key == "api_key":
+                            config["jellyfin_api_key"] = value
+
+                if "update_checker" in file_config and isinstance(file_config["update_checker"], dict):
+                    uc = file_config["update_checker"]
+                    if "enabled" in uc:
+                        config["update_checker_enabled"] = uc["enabled"]
+                    if "interval_minutes" in uc:
+                        config["update_checker_interval"] = int(uc["interval_minutes"])
+        except (yaml.YAMLError, IOError) as e:
+            logger.error(f"Error loading config file: {e}")
+
+    logger.info(f"Configuration loaded: {config}")
+    return config
+
+__all__ = ["_load_config", "logger"]
